personalportfolio/requirements/base.txt
# Core
Django>=4.2,<5.0
python-decouple>=3.8

# Database
psycopg2-binary>=2.9
dj-database-url>=2.0

# Static files
whitenoise>=6.6

# Images
Pillow>=10.0

# Testing
pytest>=7.4
pytest-django>=4.7

# Code quality
ruff>=0.3
black>=24.0
